Fix broken error counting and improve key handling
As of now, 'onChar' is only fired the first time the key gets pushed
down. Therefore we will not count multiple errors.

We now use react-keyboard-event-handler to catch global typing without
the need to focus the keyboard div.

Also we prevent default key handling, this may be unexpected if you want
to use shortcuts in your browser, but  improves the usage of the touchtypingtutor.